AEA-3444: enable api gw logging (#138)
Updates the `HttpApiGateway `resource in the main and sandbox templates
to enable encrypted access logging. Also adds a number of account wide
supporting resources in the new `account_resources.yml`, and a minor
update to the ci roles permissions in `ci_resources.yml`.

**Note:**` ci_resources` and `account_resources` will each need to be
manually deployed in each of the other environments before deploying
these changes.