AEA-3657 Use getSecrets from published artifact from electronic-prescription-service-get-secrets repo (#443)
## Summary

Use getSecrets from published artifact from
electronic-prescription-service-get-secrets repo

:ticket: [AEA-3657](https://nhsd-jira.digital.nhs.uk/browse/AEA-3657)

- Routine Change
- :exclamation: Breaking Change
- :robot: Operational or Infrastructure Change
- :sparkles: New Feature
- :warning: Potential issues that might be caused by this change

### Details

[AEA-3612](https://nhsd-jira.digital.nhs.uk/browse/AEA-3612) copies the
getSecrets layer code to a new repo and publishes it as a release
artifact. We need to change the pfp stacks so that it uses the published
version of the getSecrets layer and removes the getSecrets layer code
from pfp repo
As part of the build process for stacks in pfp, it should download from
the github release artifact. RELEASE TAG should be clearly marked in the
cloudformation template so it can easily be changed when a new version
of getSecrets layer is published

## Reviews Required

**Check who should review this. Remove this line once this has been
done**

- [x] Dev
- [ ] Test
- [ ] Tech Author
- [ ] Product Owner

## Review Checklist

:information_source: This section is to be filled in by the
**reviewer**.

- [ ] I have reviewed the changes in this PR and they fill all or part
of the acceptance criteria of the ticket, and the code is in a mergeable
state.
- [ ] If there were infrastructure, operational, or build changes, I
have made sure there is sufficient evidence that the changes will work.
- [ ] I have ensured the jira ticket has been updated with the github
pull request link